pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wyoming-chatterbox"
version = "1.0.0"
description = "Wyoming protocol server for Chatterbox TTS with voice cloning"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "sudoxnym"}
]
keywords = ["wyoming", "tts", "voice-cloning", "chatterbox", "home-assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
requires-python = ">=3.10"
dependencies = [
    "wyoming>=1.5.0",
    "chatterbox-tts",
    "torch",
    "torchaudio",
]

[project.urls]
Homepage = "https://github.com/sudoxnym/wyoming-chatterbox"
Repository = "https://github.com/sudoxnym/wyoming-chatterbox"

[project.scripts]
wyoming-chatterbox = "wyoming_chatterbox.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
